Add android demo of legend
